Coalesce into one file
