--- atspi-common/src/events/wrappers/mod.rs
@@ -1,6 +1,0 @@
-pub mod event;
-pub use event::Event;
-pub mod object;
-pub use object::ObjectEvents;
-pub mod document;
-pub use document::DocumentEvents;
